make component & fragment on hero
